evaluate_instruction_5: Deal with cast warnings

+++ src/evaluate_instruction_5.c
@@ -234,7 +234,7 @@ void isc_indirect_y(State *state, uint8_t val) {
 void rla_zero_page(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -251,7 +251,7 @@ void rla_zero_page(State *state, uint8_t val) {
 void rla_zero_page_x(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -268,7 +268,7 @@ void rla_zero_page_x(State *state, uint8_t val) {
 void rla_absolute(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -285,7 +285,7 @@ void rla_absolute(State *state, uint8_t val) {
 void rla_absolute_x(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -302,7 +302,7 @@ void rla_absolute_x(State *state, uint8_t val) {
 void rla_absolute_y(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -319,7 +319,7 @@ void rla_absolute_y(State *state, uint8_t val) {
 void rla_indirect_x(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -336,7 +336,7 @@ void rla_indirect_x(State *state, uint8_t val) {
 void rla_indirect_y(State *state, uint8_t val) {
 	// Rotate left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = (val << 1) | state->cpu.p.C;
+	uint8_t result = (uint8_t)((val << 1) | (uint8_t)state->cpu.p.C);
 	state->cpu.p.C = carry;
 
 	// AND with accumulator
@@ -353,7 +353,7 @@ void rla_indirect_y(State *state, uint8_t val) {
 void rra_zero_page(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -369,7 +369,7 @@ void rra_zero_page(State *state, uint8_t val) {
 void rra_zero_page_x(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -385,7 +385,7 @@ void rra_zero_page_x(State *state, uint8_t val) {
 void rra_absolute(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -401,7 +401,7 @@ void rra_absolute(State *state, uint8_t val) {
 void rra_absolute_x(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -417,7 +417,7 @@ void rra_absolute_x(State *state, uint8_t val) {
 void rra_absolute_y(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -433,7 +433,7 @@ void rra_absolute_y(State *state, uint8_t val) {
 void rra_indirect_x(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -449,7 +449,7 @@ void rra_indirect_x(State *state, uint8_t val) {
 void rra_indirect_y(State *state, uint8_t val) {
 	// Rotate right
 	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
-	uint8_t result = (val >> 1) | (state->cpu.p.C << 7);
+	uint8_t result = (uint8_t)((val >> 1) | ((uint8_t)state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
 	// Add with carry
@@ -465,7 +465,7 @@ void rra_indirect_y(State *state, uint8_t val) {
 void slo_zero_page(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -482,7 +482,7 @@ void slo_zero_page(State *state, uint8_t val) {
 void slo_zero_page_x(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -499,7 +499,7 @@ void slo_zero_page_x(State *state, uint8_t val) {
 void slo_absolute(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -516,7 +516,7 @@ void slo_absolute(State *state, uint8_t val) {
 void slo_absolute_x(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -533,7 +533,7 @@ void slo_absolute_x(State *state, uint8_t val) {
 void slo_absolute_y(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -550,7 +550,7 @@ void slo_absolute_y(State *state, uint8_t val) {
 void slo_indirect_x(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -567,7 +567,7 @@ void slo_indirect_x(State *state, uint8_t val) {
 void slo_indirect_y(State *state, uint8_t val) {
 	// Shift left
 	uint8_t carry  = (val & 0x80) ? 1 : 0;
-	uint8_t result = val << 1;
+	uint8_t result = (uint8_t)(val << 1);
 	state->cpu.p.C = carry;
 
 	// OR with accumulator
@@ -733,7 +733,7 @@ void arr(State *state, uint8_t val) {
 	// Shift right with carry
 	uint8_t carry = (state->cpu.a & 0x01) ? 0x80 : 0;
 	state->cpu.a >>= 1;
-	state->cpu.a |= (state->cpu.p.C << 7);
+	state->cpu.a |= ((uint8_t)state->cpu.p.C << 7);
 	state->cpu.p.C = carry;
 
 	// Update flags
@@ -781,7 +781,7 @@ void shy(State *state, uint8_t val) {
 	uint8_t addr_high = (val + 1) & 0xFF;
 
 	// Store Y register in memory
-	state_set_mem(state, addr_low | (addr_high << 8), state->cpu.y);
+	state_set_mem(state, (uint16_t)(addr_low | (addr_high << 8)), state->cpu.y);
 }
 
 void shx(State *state, uint8_t val) {
@@ -790,7 +790,7 @@ void shx(State *state, uint8_t val) {
 	uint8_t addr_high = (val + 1) & 0xFF;
 
 	// Store X register in memory
-	state_set_mem(state, addr_low | (addr_high << 8), state->cpu.x);
+	state_set_mem(state, (uint16_t)(addr_low | (addr_high << 8)), state->cpu.x);
 }
 
 void ahx_absolute_y(State *state, uint8_t val) {
@@ -799,7 +799,7 @@ void ahx_absolute_y(State *state, uint8_t val) {
 	uint8_t addr_high = (val + 1) & 0xFF;
 
 	// Store (A & X) in memory
-	state_set_mem(state, addr_low | (addr_high << 8), state->cpu.a & state->cpu.x);
+	state_set_mem(state, (uint16_t)(addr_low | (addr_high << 8)), state->cpu.a & state->cpu.x);
 }
 
 void ahx_indirect_y(State *state, uint8_t val) {
@@ -808,5 +808,5 @@ void ahx_indirect_y(State *state, uint8_t val) {
 	uint8_t addr_high = (val + 1) & 0xFF;
 
 	// Store (A & X) in memory
-	state_set_mem(state, addr_low | (addr_high << 8), state->cpu.a & state->cpu.x);
+	state_set_mem(state, (uint16_t)(addr_low | (addr_high << 8)), state->cpu.a & state->cpu.x);
 }
